agentic_trading_system/requirements.txt
redis==5.0.3
pydantic==2.6.4
google-genai==0.3.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
websockets>=13.0
pandas==2.2.1
ta==0.11.0
requests==2.32.3
openai==1.107.2
